# apps/web_builder/wb_backend/wb_api_services/tools/swagger.json
{"swagger":"2.0","info":{"title":"API Documentation","version":"1.0.0","description":"API documentation for kis-serverless service"},"paths":{"/login":{"post":{"summary":"login","description":"User login","tags":["User Authentication"],"operationId":"postLogin","consumes":["application/json"],"produces":["application/json"],"security":[{"Authorization":[]}],"parameters":[{"in":"body","name":"body","description":"Body required in the request","required":true,"schema":{"$ref":"#/definitions/LoginDAO"}}],"responses":{"200":{"description":"Login successful","schema":{"$ref":"#/definitions/LoginResponseDAO"}},"400":{"description":"Username or password is invalid!"},"404":{"description":"User not found!"},"500":{"description":"Server error"}}}},"/refresh":{"post":{"summary":"login","description":"Refresh token","tags":["User Authentication"],"operationId":"postRefresh","consumes":["application/json"],"produces":["application/json"],"security":[{"Authorization":[]}],"parameters":[{"in":"body","name":"body","description":"Body required in the request","required":true,"schema":{"$ref":"#/definitions/RefreshDAO"}}],"responses":{"200":{"description":"Token refreshed successfully","schema":{"$ref":"#/definitions/RefreshResponseDAO"}},"400":{"description":"Invalid refresh token"},"500":{"description":"Server error"}}}},"/register":{"post":{"summary":"register","description":"Register new user","tags":["User Authentication"],"operationId":"postRegister","consumes":["application/json"],"produces":["application/json"],"security":[{"Authorization":[]}],"parameters":[{"in":"body","name":"body","description":"Body required in the request","required":true,"schema":{"$ref":"#/definitions/RegisterDAO"}}],"responses":{"201":{"description":"User registered successfully","schema":{"$ref":"#/definitions/RegisterResponseDAO"}},"400":{"description":"Invalid input data"},"500":{"description":"Server error"}}}},"/verify-token":{"post":{"summary":"verify-token","description":"Verify user token","tags":["User Authentication"],"operationId":"postVerifyToken","consumes":["application/json"],"produces":["application/json"],"security":[{"Authorization":[]}],"parameters":[],"responses":{"200":{"description":"Token valid"},"400":{"description":"Invalid token"},"500":{"description":"Server error"}}}},"/logout":{"post":{"summary":"logout","description":"Logout user","tags":["User Authentication"],"operationId":"postLogout","consumes":["application/json"],"produces":["application/json"],"security":[{"Authorization":[]}],"parameters":[],"responses":{"200":{"description":"Logged out successfully"},"400":{"description":"Token required"}}}},"/profile/{userId}":{"get":{"summary":"get-users-by-id","description":"","tags":["User Management"],"operationId":"getUserById","consumes":["application/json"],"produces":["application/json"],"security":[{"Authorization":[]}],"parameters":[{"name":"userId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"User details retrieved successfully","schema":{"$ref":"#/definitions/UserDAO"}},"400":{"description":"Token required"},"404":{"description":"User not found"},"500":{"description":"Could not retrieve user information"}}},"put":{"summary":"update-user","description":"","tags":["User Management"],"operationId":"putUpdateUser","consumes":["application/json"],"produces":["application/json"],"security":[{"Authorization":[]}],"parameters":[{"name":"userId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"User updated successfully","schema":{"$ref":"#/definitions/UserDAO"}},"400":{"description":"Token required"},"500":{"description":"Could not update password"}}}},"/users":{"get":{"summary":"get-users","description":"","tags":["User Management"],"operationId":"getUsers","consumes":["application/json"],"produces":["application/json"],"security":[{"Authorization":[]}],"parameters":[],"responses":{"200":{"description":"Users retrieved successfully","schema":{"$ref":"#/definitions/UserDAO"}},"404":{"description":"No users found"},"500":{"description":"Could not retrieve users information"}}}},"/deleted-users":{"get":{"summary":"get-delete-users","description":"","tags":["User Management"],"operationId":"getDeletedUsers","consumes":["application/json"],"produces":["application/json"],"security":[{"Authorization":[]}],"parameters":[],"responses":{"200":{"description":"Deleted users retrieved successfully","schema":{"$ref":"#/definitions/UserDAO"}},"404":{"description":"No deleted users found"},"500":{"description":"Could not retrieve deleted users information"}}}},"/create-new-project":{"post":{"summary":"create-new-project","description":"Create a new project","tags":["Project Management"],"operationId":"postCreateNewProject","consumes":["application/json"],"produces":["application/json"],"security":[{"Authorization":[]}],"parameters":[{"in":"body","name":"body","description":"Body required in the request","required":true,"schema":{"$ref":"#/definitions/NewProjectDAO"}}],"responses":{"200":{"description":"Project created successfully","schema":{"$ref":"#/definitions/ProjectDAO"}},"400":{"description":"Invalid project data"},"500":{"description":"Server error"}}}},"/projects":{"get":{"summary":"get-projects","description":"Get all projects","tags":["Project Management"],"operationId":"getProjects","consumes":["application/json"],"produces":["application/json"],"security":[{"Authorization":[]}],"parameters":[],"responses":{"200":{"description":"List of projects","schema":{"$ref":"#/definitions/ProjectDAO"}},"400":{"description":"Token required"},"403":{"description":"Invalid token"},"500":{"description":"Server error"}}}},"/project-detail/{projectId}":{"get":{"summary":"get-project-detail","description":"Get project details","tags":["Project Management"],"operationId":"getProjectDetail","consumes":["application/json"],"produces":["application/json"],"security":[{"Authorization":[]}],"parameters":[{"name":"projectId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Project details retrieved","schema":{"$ref":"#/definitions/ProjectDAO"}},"400":{"description":"Token required"},"404":{"description":"Project not found"},"500":{"description":"Server error"}}},"put":{"summary":"update-project","description":"Update project details","tags":["Project Management"],"operationId":"putUpdateProject","consumes":["application/json"],"produces":["application/json"],"security":[{"Authorization":[]}],"parameters":[{"name":"projectId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Project updated successfully","schema":{"$ref":"#/definitions/ProjectDAO"}},"400":{"description":"Invalid project data"},"404":{"description":"Project not found"},"500":{"description":"Server error"}}}},"/page-detail/{alias}/{url_alias}":{"get":{"summary":"get-page-detail","description":"Get page details","tags":["Project Management"],"operationId":"getPageDetails","consumes":["application/json"],"produces":["application/json"],"security":[{"Authorization":[]}],"parameters":[{"name":"alias","in":"path","required":true,"type":"string"},{"name":"url_alias","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Page details retrieved","schema":{"$ref":"#/definitions/PageDAO"}},"400":{"description":"Token required"},"404":{"description":"Page not found"},"500":{"description":"Server error"}}}},"/projects/{projectId}":{"delete":{"summary":"delete-project","description":"Delete project","tags":["Project Management"],"operationId":"deleteProject","consumes":["application/json"],"produces":["application/json"],"security":[{"Authorization":[]}],"parameters":[{"name":"projectId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Project deleted successfully"},"404":{"description":"Project not found"},"500":{"description":"Server error"}}}},"/projects/{projectId}/duplicate":{"post":{"summary":"duplicate-project","description":"Duplicate a project","tags":["Project Management"],"operationId":"postDuplicateProject","consumes":["application/json"],"produces":["application/json"],"security":[{"Authorization":[]}],"parameters":[{"name":"projectId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Project duplicated successfully","schema":{"$ref":"#/definitions/ProjectDAO"}},"404":{"description":"Project not found"},"500":{"description":"Server error"}}}},"/projects/{alias}/create-new-page":{"post":{"summary":"create-new-page","description":"Create a new page","tags":["Page Management"],"operationId":"postCreateNewPage","consumes":["application/json"],"produces":["application/json"],"security":[{"Authorization":[]}],"parameters":[{"in":"body","name":"body","description":"Body required in the request","required":true,"schema":{"$ref":"#/definitions/NewPageDAO"}},{"name":"alias","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Page created successfully","schema":{"$ref":"#/definitions/PageDAO"}},"400":{"description":"Invalid page data"},"500":{"description":"Server error"}}}},"/projects-detail/{alias}":{"get":{"summary":"get-projects-detail-by-alias","description":"Get all projects by alias","tags":["Page Management"],"operationId":"getProjectDetailsByAlias","consumes":["application/json"],"produces":["application/json"],"security":[{"Authorization":[]}],"parameters":[{"name":"alias","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"List of projects by alias","schema":{"$ref":"#/definitions/ProjectDAO"}},"500":{"description":"Server error"}}}},"/page-detail/{alias}":{"get":{"summary":"get-pages","description":"Get all pages","tags":["Page Management"],"operationId":"getPages","consumes":["application/json"],"produces":["application/json"],"security":[{"Authorization":[]}],"parameters":[{"name":"alias","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"List of pages","schema":{"$ref":"#/definitions/PageDAO"}},"500":{"description":"Server error"}}}},"/projects/{alias}/{url_alias}/update":{"put":{"summary":"update-page","description":"Update page details","tags":["Page Management"],"operationId":"updatePage","consumes":["application/json"],"produces":["application/json"],"security":[{"Authorization":[]}],"parameters":[{"name":"alias","in":"path","required":true,"type":"string"},{"name":"url_alias","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Page updated successfully","schema":{"$ref":"#/definitions/PageDAO"}},"400":{"description":"Token required"},"404":{"description":"Page not found"},"500":{"description":"Could not update page"}}}},"/projects/{alias}/{url_alias}":{"delete":{"summary":"delete-page","description":"Delete page","tags":["Page Management"],"operationId":"deletePage","consumes":["application/json"],"produces":["application/json"],"security":[{"Authorization":[]}],"parameters":[{"name":"alias","in":"path","required":true,"type":"string"},{"name":"url_alias","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Page deleted successfully"},"400":{"description":"Token required"},"404":{"description":"Page not found"},"500":{"description":"Could not delete page"}}}},"/upload-image":{"post":{"summary":"upload-image","description":"","tags":["Image Management"],"operationId":"postUploadImage","consumes":["application/json"],"produces":["application/json"],"security":[{"Authorization":[]}],"parameters":[{"in":"body","name":"body","description":"Body required in the request","required":true,"schema":{"$ref":"#/definitions/ImageDAO"}}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/ImageResponseDAO"}},"500":{"description":"Error saving image data to the database"}}}},"/delete-users/{userId}":{"delete":{"summary":"delete-user","description":"Delete a user by ID","tags":["User Management"],"operationId":"deleteUser","consumes":["application/json"],"produces":["application/json"],"security":[{"Authorization":[]}],"parameters":[{"name":"userId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"User deleted successfully"},"400":{"description":"Invalid user ID"},"404":{"description":"User not found"},"500":{"description":"Server error"}}}},"/restore-user/{userId}":{"put":{"summary":"restore-user","description":"Restore a soft-deleted user","tags":["User Management"],"operationId":"restoreUser","consumes":["application/json"],"produces":["application/json"],"security":[{"Authorization":[]}],"parameters":[{"name":"userId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"User restored successfully"},"400":{"description":"Invalid user ID or user already active"},"404":{"description":"User not found"},"500":{"description":"Server error"}}}}},"definitions":{"NewPageDAO":{"type":"object","properties":{"url_alias":{"type":"string"},"title":{"type":"string"},"metaTags":{"type":"string"},"sections":{"type":"JSON"}}},"NewProjectDAO":{"type":"object","properties":{"name":{"type":"string"},"alias":{"type":"string"},"description":{"type":"string"},"thumbnail":{"type":"string"}}},"PageDAO":{"type":"object","properties":{"id":{"type":"number"},"url_alias":{"type":"string"},"title":{"type":"string"},"meta_tags":{"type":"string"},"sections":{"type":"JSON"},"project_id":{"type":"number"}}},"UserDAO":{"type":"object","properties":{"userId":{"type":"number"},"username":{"type":"string"}}},"LoginResponseDAO":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"userId":{"type":"number"},"username":{"type":"string"}}},"LoginDAO":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"RefreshResponseDAO":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"}}},"RefreshDAO":{"type":"object","properties":{"refreshToken":{"type":"string"}}},"ProjectDAO":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"description":{"type":"string"},"created_at":{"type":"string"},"updated_at":{"type":"string"},"thumbnail":{"type":"string"},"alias":{"type":"string"}}},"RegisterResponseDAO":{"type":"object","properties":{"newUserId":{"type":"number"},"newUsername":{"type":"string"}}},"RegisterDAO":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"ImageResponseDAO":{"type":"object","properties":{"signedUrl":{"type":"string"},"shortUrl":{"type":"string"},"fileName":{"type":"string"},"imageId":{"type":"number"}}},"ImageDAO":{"type":"object","properties":{"fileName":{"type":"string"},"fileType":{"type":"string"},"signedUrl":{"type":"string"}}}},"securityDefinitions":{"Authorization":{"type":"apiKey","name":"Authorization","in":"header"}},"basePath":"/dev","host":"localhost:4000"}